refactor : InputView가 파싱된 값을 넘겨주도록 수정
